backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
